сделал авторизацию + подогнал условия в сервисах

+++ internal/services/model.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 
 	"github.com/smakkking/avito_test/internal/models"
+	"github.com/smakkking/avito_test/internal/services/utils"
 )
 
 type Service struct {
@@ -31,21 +32,35 @@ func NewService(storage Storage) *Service {
 }
 
 var (
-	ErrNotFound = errors.New("banner not found")
+	ErrNotFound   = errors.New("banner not found")
+	ErrNotAllowed = errors.New("not allowed")
 )
 
 func (s *Service) DeleteBanner(ctx context.Context, bannerID int) (bool, error) {
+	if !utils.IsAdmin(ctx) {
+		return false, ErrNotAllowed
+	}
+
 	return s.bannerStorage.DeleteBanner(ctx, bannerID)
 }
 
 func (s *Service) UpdateBanner(ctx context.Context, bannerID int, banner *models.BasicBannnerInfo) (bool, error) {
+	if !utils.IsAdmin(ctx) {
+		return false, ErrNotAllowed
+	}
+
 	return s.bannerStorage.UpdateBanner(ctx, bannerID, banner)
 }
 
 func (s *Service) GetUserBanner(ctx context.Context, tagID int, featureID int, useLastRevision bool) (interface{}, error) {
-	banner, _, err := s.bannerStorage.GetUserBanner(ctx, tagID, featureID)
+	banner, enabled, err := s.bannerStorage.GetUserBanner(ctx, tagID, featureID)
+
 	if err != nil {
 		return nil, err
+	}
+
+	if !utils.IsAdmin(ctx) && !enabled {
+		return nil, ErrNotAllowed
 	}
 
 	return banner, nil
@@ -57,6 +72,10 @@ func (s *Service) GetAllBannersFiltered(
 	featureID int, featureSearch bool,
 	limit int, offset int,
 ) ([]*models.BannerInfo, error) {
+	if !utils.IsAdmin(ctx) {
+		return nil, ErrNotAllowed
+	}
+
 	banners, err := s.bannerStorage.GetAllBannersFiltered(ctx, tagID, tagSearch, featureID, featureSearch, limit, offset)
 	if err != nil {
 		return nil, err
@@ -66,6 +85,10 @@ func (s *Service) GetAllBannersFiltered(
 }
 
 func (s *Service) CreateBanner(ctx context.Context, banner *models.BasicBannnerInfo) (int, error) {
+	if !utils.IsAdmin(ctx) {
+		return 0, ErrNotAllowed
+	}
+
 	bannerID, err := s.bannerStorage.CreateBanner(ctx, banner)
 	if err != nil {
 		return 0, err
